create readme.md
In this thesis, mathematical and numerical models are developed and a computer program for calculating the mineral extraction process by In-situ leaching is developed. The hydrodynamic pressure distribution, velocity field, and reagent concentration distributions in the liquid phase are obtained. UC in the liquid and solid phase.
By varying the initial input data such as the location, depth, and flow rate of wells, and getting the corresponding results, you can analyze and forecast mineral production.
The developed model can easily be applied to modeling the leaching of other minerals.
We used parallel programming MPI technologies, where processes interact via message passing. The parallel computing algorithm was tested by running with different numbers of interacting processes.
Based on the results obtained, it can be concluded that tasks that require a large amount of memory, for the speed of calculations, are acceptable to solve on high-performance devices.
In this thesis, the initial physical area of the calculation was described by a grid of 360 × 180×180 points. For a given number of points, as can be seen from the graph of the dependence of the calculation time on the number of processes, 35-40 processors are sufficient. If the problem requires a more detailed description in terms of space and time, i.e. the number of points in the source domain is orders of magnitude greater than the one considered in this thesis, then the number of necessary processes will increase accordingly. Such tasks are encountered by every modern researcher and developer, and the use of parallel computing methods on high-performance systems will facilitate and accelerate the obtaining of final results, which is important in any modern science.